+++ vasp_step/energy_parameters.py
@@ -86,6 +86,20 @@ class EnergyParameters(seamm.Parameters):  # noqa: E999
             "description": "Write the input files and stop:",
             "help_text": "Don't run VASP. Just write the input files.",
         },
+        "use hdf5 files": {
+            "default": "no",
+            "kind": "boolean",
+            "default_units": "",
+            "enumeration": (
+                "yes",
+                "no",
+            ),
+            "format_string": "s",
+            "description": "Output to  hdf5 files:",
+            "help_text": (
+                "Use hdf5 files for the wavefunction and charge density output."
+            ),
+        },
         "calculate stress": {
             "default": "yes",
             "kind": "enumeration",
@@ -123,7 +137,7 @@ class EnergyParameters(seamm.Parameters):  # noqa: E999
             ),
         },
         "ediff": {
-            "default": 1.0e-6,
+            "default": 1.0e-5,
             "kind": "float",
             "default_units": "",
             "enumeration": None,
@@ -146,6 +160,15 @@ class EnergyParameters(seamm.Parameters):  # noqa: E999
             "format_string": "",
             "description": "Electronic minimization method:",
             "help_text": "The algorithm used tp minimize the electronic energy.",
+        },
+        "precision": {
+            "default": "normal",
+            "kind": "enumeration",
+            "default_units": "",
+            "enumeration": ("single", "normal", "accurate"),
+            "format_string": "",
+            "description": "Precision of FFT grids and projections:",
+            "help_text": "The precision of the FFT grids the real-space projections.",
         },
         # Performance
         "np": {
@@ -353,7 +376,7 @@ class EnergyParameters(seamm.Parameters):  # noqa: E999
             "help_text": "The maximum energy cutoff in the specified potentials.",
         },
         "spin polarization": {
-            "default": "collinear",
+            "default": "none",
             "kind": "string",
             "default_units": "",
             "enumeration": ("none", "collinear", "noncollinear"),