+++ HISTORY.rst
@@ -1,7 +1,13 @@
 =======
 History
 =======
-
+2025.11.2: Small improvements and better output.
+    * Added missing precision control
+    * Added control over using HDF5 files
+    * Improved the output, particularly for optimization
+    * Standardized pressure and stress units to GPa
+    * Changed some of the defaults to make standard calculations work better.
+      
 2025.10.31: Added optimization and improved the output.
 
 2025.10.22: First working version